requirements.txt
openai==0.27.8
pyqt5
pyperclip
pynput
pytesseract
pillow